Made SidebarStatusSelector.tsx selector more generic + fixed label inconsistencies
